Add thread-based output to Browser from protocol
